忘了

+++ src/types.h
@@ -4,7 +4,7 @@ typedef unsigned char *POINTER;
 typedef unsigned char BYTE;
 typedef unsigned char UCHAR;
 typedef unsigned short int WORD;
-typedef long LONG;
+typedef int LONG;
 typedef unsigned int DWORD;
 typedef unsigned int UINT4;
 #endif
